pyproject.toml
[project]
name = "python-readability"
dynamic = ["version"]
description = "@mozilla/readability for Python"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
]

[project.optional-dependencies]
pythonmonkey = [
    "pythonmonkey-fork~=1.0 ; sys_platform != 'emscripten'",
]
mini-racer = [
    "mini-racer~=0.12 ; sys_platform != 'emscripten'",
]
stpyv8 = [
    "stpyv8~=13.1 ; sys_platform != 'emscripten'",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.uv]
dev-dependencies = [
    "pyodide-py",
    "python-readability[mini-racer,pythonmonkey,stpyv8]",
    "ruff~=0.8.0",
]

[tool.ruff.lint]
extend-select = [
    "F",    # flake8
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle
    "UP",   # pyupgrade
    "TRY",  # tryceratops
    "RUF",  # ruff
    "FURB", # refurb
    "C4",   # flake8-comprehensions
    "ARG",  # flake8-unused-arguments
    "PIE",  # flake8-pie
    "PTH",  # flake8-use-pathlib
    "RSE",  # flake8-raise
    "SIM",  # flake8-simplify
    "SLF",  # flake8-self
]

[tool.pdm.version]
source = "file"
path = "src/readability/__init__.py"

[tool.ruff]
line-length = 200
